+++ sdk/servicebus/Azure.Messaging.ServiceBus/src/Amqp/AmqpReceiver.cs
@@ -7,6 +7,7 @@ using System.Diagnostics;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using System.Transactions;
 using Azure.Core;
 using Azure.Messaging.ServiceBus.Core;
 using Azure.Messaging.ServiceBus.Diagnostics;
@@ -58,7 +59,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
         /// The AMQP connection scope responsible for managing transport constructs for this instance.
         /// </summary>
         ///
-        private AmqpConnectionScope ConnectionScope { get; }
+        private readonly AmqpConnectionScope _connectionScope;
 
         /// <summary>
         /// The <see cref="ReceiveMode"/> used to specify how messages are received. Defaults to PeekLock mode.
@@ -124,7 +125,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
             Argument.AssertNotNull(retryPolicy, nameof(retryPolicy));
 
             _entityPath = entityPath;
-            ConnectionScope = connectionScope;
+            _connectionScope = connectionScope;
             _retryPolicy = retryPolicy;
             _isSessionReceiver = isSessionReceiver;
             _receiveMode = receiveMode;
@@ -133,7 +134,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
 
             _receiveLink = new FaultTolerantAmqpObject<ReceivingAmqpLink>(
                 timeout =>
-                    ConnectionScope.OpenReceiverLinkAsync(
+                    _connectionScope.OpenReceiverLinkAsync(
                         entityPath: _entityPath,
                         timeout: timeout,
                         prefetchCount: prefetchCount,
@@ -147,7 +148,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
                 });
 
             _managementLink = new FaultTolerantAmqpObject<RequestResponseAmqpLink>(
-                timeout => ConnectionScope.OpenManagementLinkAsync(
+                timeout => _connectionScope.OpenManagementLinkAsync(
                     _entityPath,
                     timeout,
                     CancellationToken.None),
@@ -187,7 +188,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
                     timeout,
                     cancellationToken).ConfigureAwait(false);
             },
-            ConnectionScope,
+            _connectionScope,
             cancellationToken).ConfigureAwait(false);
 
             return messages;
@@ -217,7 +218,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
 
             var stopWatch = Stopwatch.StartNew();
 
-            link = await _receiveLink.GetOrCreateAsync(UseMinimum(ConnectionScope.SessionTimeout, timeout)).ConfigureAwait(false);
+            link = await _receiveLink.GetOrCreateAsync(UseMinimum(_connectionScope.SessionTimeout, timeout)).ConfigureAwait(false);
             cancellationToken.ThrowIfCancellationRequested<TaskCanceledException>();
 
             var messagesReceived = await Task.Factory.FromAsync
@@ -275,7 +276,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
                 await CompleteInternalAsync(
                     lockTokens,
                     timeout).ConfigureAwait(false),
-                ConnectionScope,
+                _connectionScope,
                 cancellationToken).ConfigureAwait(false);
 
         /// <summary>
@@ -288,16 +289,18 @@ namespace Azure.Messaging.ServiceBus.Amqp
             IEnumerable<string> lockTokens,
             TimeSpan timeout)
         {
-            Guid[] lockTokenGuids = lockTokens.Select(token => new Guid(token)).ToArray();
-            if (lockTokenGuids.Any(lockToken => _requestResponseLockedMessages.Contains(lockToken)))
+            var lockTokenGuids = lockTokens.Select(token => new Guid(token)).ToArray();
+            foreach (var tokenGuid in lockTokenGuids)
             {
-                await DisposeMessageRequestResponseAsync(
-                    lockTokenGuids,
-                    timeout,
-                    DispositionStatus.Completed,
-                    _isSessionReceiver,
-                    SessionId).ConfigureAwait(false);
-                return;
+                if (_requestResponseLockedMessages.Contains(tokenGuid))
+                {
+                    await DisposeMessageRequestResponseAsync(
+                        lockTokenGuids,
+                        timeout,
+                        DispositionStatus.Completed,
+                        SessionId).ConfigureAwait(false);
+                    return;
+                }
             }
             await DisposeMessagesAsync(lockTokenGuids, AmqpConstants.AcceptedOutcome, timeout).ConfigureAwait(false);
         }
@@ -325,11 +328,14 @@ namespace Azure.Messaging.ServiceBus.Amqp
             try
             {
                 ArraySegment<byte> transactionId = AmqpConstants.NullBinary;
-                //var ambientTransaction = Transaction.Current;
-                //if (ambientTransaction != null)
-                //{
-                //    transactionId = await AmqpTransactionManager.Instance.EnlistAsync(ambientTransaction, ServiceBusConnection).ConfigureAwait(false);
-                //}
+                Transaction ambientTransaction = Transaction.Current;
+                if (ambientTransaction != null)
+                {
+                    transactionId = await AmqpTransactionManager.Instance.EnlistAsync(
+                        ambientTransaction,
+                        _connectionScope,
+                        timeout).ConfigureAwait(false);
+                }
 
                 if (!_receiveLink.TryGetOpenedObject(out receiveLink))
                 {
@@ -419,7 +425,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
                     lockToken,
                     timeout,
                     propertiesToModify).ConfigureAwait(false),
-                ConnectionScope,
+                _connectionScope,
                 cancellationToken).ConfigureAwait(false);
 
         /// <summary>Indicates that the receiver wants to defer the processing for the message.</summary>
@@ -433,18 +439,18 @@ namespace Azure.Messaging.ServiceBus.Amqp
             TimeSpan timeout,
             IDictionary<string, object> propertiesToModify = null)
         {
-            Guid[] lockTokens = new[] { new Guid(lockToken) };
-            if (lockTokens.Any(lt => _requestResponseLockedMessages.Contains(lt)))
+            Guid lockTokenGuid = new Guid(lockToken);
+            var lockTokenGuids = new[] { lockTokenGuid };
+            if (_requestResponseLockedMessages.Contains(lockTokenGuid))
             {
                 return DisposeMessageRequestResponseAsync(
-                    lockTokens,
+                    lockTokenGuids,
                     timeout,
                     DispositionStatus.Defered,
-                    _isSessionReceiver,
                     SessionId,
                     propertiesToModify);
             }
-            return DisposeMessagesAsync(lockTokens, GetDeferOutcome(propertiesToModify), timeout);
+            return DisposeMessagesAsync(lockTokenGuids, GetDeferOutcome(propertiesToModify), timeout);
         }
 
         /// <summary>
@@ -471,7 +477,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
                     lockToken,
                     timeout,
                     propertiesToModify).ConfigureAwait(false),
-                ConnectionScope,
+                _connectionScope,
                 cancellationToken).ConfigureAwait(false);
 
         /// <summary>
@@ -486,18 +492,18 @@ namespace Azure.Messaging.ServiceBus.Amqp
             TimeSpan timeout,
             IDictionary<string, object> propertiesToModify = null)
         {
-            Guid[] lockTokens = new[] { new Guid(lockToken) };
-            if (lockTokens.Any(lt => _requestResponseLockedMessages.Contains(lt)))
+            Guid lockTokenGuid = new Guid(lockToken);
+            var lockTokenGuids = new[] { lockTokenGuid };
+            if (_requestResponseLockedMessages.Contains(lockTokenGuid))
             {
                 return DisposeMessageRequestResponseAsync(
-                    lockTokens,
+                    lockTokenGuids,
                     timeout,
                     DispositionStatus.Abandoned,
-                    _isSessionReceiver,
                     SessionId,
                     propertiesToModify);
             }
-            return DisposeMessagesAsync(lockTokens, GetAbandonOutcome(propertiesToModify), timeout);
+            return DisposeMessagesAsync(lockTokenGuids, GetAbandonOutcome(propertiesToModify), timeout);
         }
 
         /// <summary>
@@ -532,7 +538,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
                     propertiesToModify,
                     deadLetterReason,
                     deadLetterErrorDescription).ConfigureAwait(false),
-                ConnectionScope,
+                _connectionScope,
                 cancellationToken).ConfigureAwait(false);
 
         /// <summary>
@@ -561,14 +567,14 @@ namespace Azure.Messaging.ServiceBus.Amqp
                 throw new ArgumentOutOfRangeException(nameof(deadLetterErrorDescription), string.Format(Resources.MaxPermittedLengthExceeded, Constants.MaxDeadLetterReasonLength));
             }
 
-            var lockTokens = new[] { new Guid(lockToken) };
-            if (lockTokens.Any(lt => _requestResponseLockedMessages.Contains(lt)))
+            Guid lockTokenGuid = new Guid(lockToken);
+            var lockTokenGuids = new[] { lockTokenGuid };
+            if (_requestResponseLockedMessages.Contains(lockTokenGuid))
             {
                 return DisposeMessageRequestResponseAsync(
-                    lockTokens,
+                    lockTokenGuids,
                     timeout,
                     DispositionStatus.Suspended,
-                    _isSessionReceiver,
                     SessionId,
                     propertiesToModify,
                     deadLetterReason,
@@ -576,7 +582,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
             }
 
             return DisposeMessagesAsync(
-                lockTokens,
+                lockTokenGuids,
                 GetRejectedOutcome(propertiesToModify, deadLetterReason, deadLetterErrorDescription),
                 timeout);
         }
@@ -626,7 +632,6 @@ namespace Azure.Messaging.ServiceBus.Amqp
         /// <param name="lockTokens">Message lock tokens to update disposition status.</param>
         /// <param name="timeout"></param>
         /// <param name="dispositionStatus"></param>
-        /// <param name="isSessionReceiver"></param>
         /// <param name="sessionId"></param>
         /// <param name="propertiesToModify"></param>
         /// <param name="deadLetterReason"></param>
@@ -635,7 +640,6 @@ namespace Azure.Messaging.ServiceBus.Amqp
             Guid[] lockTokens,
             TimeSpan timeout,
             DispositionStatus dispositionStatus,
-            bool isSessionReceiver,
             string sessionId = null,
             IDictionary<string, object> propertiesToModify = null,
             string deadLetterReason = null,
@@ -771,7 +775,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
                     timeout,
                     cancellationToken)
                 .ConfigureAwait(false),
-                ConnectionScope,
+                _connectionScope,
                 cancellationToken).ConfigureAwait(false);
 
             return messages;
@@ -813,7 +817,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
             }
 
             RequestResponseAmqpLink link = await _managementLink.GetOrCreateAsync(
-                UseMinimum(ConnectionScope.SessionTimeout,
+                UseMinimum(_connectionScope.SessionTimeout,
                 timeout.CalculateRemaining(stopWatch.Elapsed)))
                 .ConfigureAwait(false);
             cancellationToken.ThrowIfCancellationRequested<TaskCanceledException>();
@@ -883,7 +887,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
                         lockToken,
                         timeout).ConfigureAwait(false);
                 },
-                ConnectionScope,
+                _connectionScope,
                 cancellationToken).ConfigureAwait(false);
             return lockedUntil;
         }
@@ -920,8 +924,8 @@ namespace Azure.Messaging.ServiceBus.Amqp
 
             if (amqpResponseMessage.StatusCode == AmqpResponseStatusCode.OK)
             {
-                IEnumerable<DateTime> lockedUntilUtcTimes = amqpResponseMessage.GetValue<IEnumerable<DateTime>>(ManagementConstants.Properties.Expirations);
-                lockedUntil = lockedUntilUtcTimes.First();
+                DateTime[] lockedUntilUtcTimes = amqpResponseMessage.GetValue<DateTime[]>(ManagementConstants.Properties.Expirations);
+                lockedUntil = lockedUntilUtcTimes[0];
             }
             else
             {
@@ -935,6 +939,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
         {
             ThrowIfSessionLockLost();
             AmqpResponseMessage amqpResponseMessage = await ManagementUtilities.ExecuteRequestResponseAsync(
+                _connectionScope,
                 _managementLink,
                 amqpRequestMessage,
                 timeout).ConfigureAwait(false);
@@ -953,7 +958,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
                     lockedUntil = await RenewSessionLockInternal(
                         timeout).ConfigureAwait(false);
                 },
-                ConnectionScope,
+                _connectionScope,
                 cancellationToken).ConfigureAwait(false);
             SessionLockedUntil = lockedUntil;
         }
@@ -1001,7 +1006,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
                 {
                     sessionState = await GetStateInternal(timeout).ConfigureAwait(false);
                 },
-                ConnectionScope,
+                _connectionScope,
                 cancellationToken).ConfigureAwait(false);
             return sessionState;
         }
@@ -1056,7 +1061,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
                         sessionState,
                         timeout).ConfigureAwait(false);
                 },
-                ConnectionScope,
+                _connectionScope,
                 cancellationToken).ConfigureAwait(false);
         }
 
@@ -1107,7 +1112,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
                 async (timeout) => messages = await ReceiveDeferredMessagesAsyncInternal(
                     sequenceNumbers.ToArray(),
                     timeout).ConfigureAwait(false),
-                ConnectionScope,
+                _connectionScope,
                 cancellationToken).ConfigureAwait(false);
             return messages;
         }
@@ -1237,7 +1242,7 @@ namespace Azure.Messaging.ServiceBus.Amqp
             await _retryPolicy.RunOperation(
                async (timeout) =>
                link = await _receiveLink.GetOrCreateAsync(timeout).ConfigureAwait(false),
-               ConnectionScope,
+               _connectionScope,
                cancellationToken).ConfigureAwait(false);
 
             if (_isSessionReceiver)