+++ sdk/servicebus/Azure.Messaging.ServiceBus/src/ServiceBusRetryPolicy.cs
@@ -119,7 +119,7 @@ namespace Azure.Messaging.ServiceBus
                 {
                     // We are in a server busy state before we start processing.
                     // Since ServerBusyBaseSleepTime > remaining time for the operation, we don't wait for the entire Sleep time.
-                    await Task.Delay(tryTimeout).ConfigureAwait(false);
+                    await Task.Delay(tryTimeout, cancellationToken).ConfigureAwait(false);
                     throw new ServiceBusException(
                         ServerBusyExceptionMessage,
                         ServiceBusException.FailureReason.ServiceBusy);
@@ -128,7 +128,7 @@ namespace Azure.Messaging.ServiceBus
                 {
                     if (IsServerBusy)
                     {
-                        await Task.Delay(ServerBusyBaseSleepTime).ConfigureAwait(false);
+                        await Task.Delay(ServerBusyBaseSleepTime, cancellationToken).ConfigureAwait(false);
                     }
 
                     try