+++ sdk/core/Azure.Core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.3.0-preview.1 (Unreleased)
 
+- Read client request ID value used for logging and tracing off the initial request object if available.
 
 ## 1.2.0 (2020-04-03)